refactor(ui): separate player stats renderer
