+++ src/components/player/index.ts
@@ -1,2 +1,3 @@
 export { default as PlayerBadge } from "./player-badge"
 export { default as PlayerVerified } from "./player-verified"
+export { default as PlayerStatsRenderer, PlayerStatsRendererSkeleton } from "./player-stats-renderer"
